Add Gatekeeper
Gatekeeper allows you to screen user joins, and bounce them (kick them immediately) from the server if certain criteria are met. This criteria can be customized by the Dogbot Moderators.

You can also specify a message to send to users right before being bounced. When enabling Gatekeeper, the channel in which you invoked the command in is set as the "broadcast channel". In this channel, user joins and user bounces will be sent here.