Implement auth state busting mechanism
We only update our stored authentication state if we don't have one
stored. So if the user has logged in previously, they will never get a
new authentication state.

Including a #auth_bust hash to force a /auth/user request fixes this --
ensuring that the stored authentication state is always updated upon a
login or reload.

This still doesn't account for random 401s, however. We'll keep the
authentication state stored until a logout is forced. This will have to
be fixed later.