Make auth state determining cheaper
When the user authenticates, the bearer token and user data is stored
in the session. Now, /auth/session/state returns this session data. It
no longer asks Discord for fresh user data.

The dashboard will use this data to determine the identity of the user.
It will GET that route upon every load and update its cached auth state
accordingly.