Redesign d?gk status embed
Invert the color scheme. If Gatekeeper is on, the embed will be
colored green instead of red. Red signifies danger, so it should only be
red when Gatekeeper is off (which can be a possible danger).

Use "on" and "off" instead of "enabled" and "disabled". Shorter and
better.

Add two newlines before the dashboard link.