Upgrade Quart and fix server creation
Upgrade Quart to 0.8.1 and change that the way the server is started.
hypercorn.asyncio.run cannot be used here because it sets the server up
while assuming that it's the only application running. It modifies the
event loop in ways that aren't desirable, so we manually create the
server ourselves.