Make config lint errors prettier
Add a new "warning" <Notice/> mood so that there can be a visual
difference between save errors and warnings. In this case, YAML syntax
errors and schema errors are presented as "warnings", and errors
returned by the server are presented as "errors".

Tweaked the prefixes of error messages a bit to seem more friendly.